recommendation/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pymongo==4.6.1
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.0 